Move ptcop converter to its own file
