+++ source/Filer.h
@@ -2,7 +2,4 @@ char GetFileNameSave(HWND hwnd,char *title);
 char GetFileNameMIDI(HWND hwnd,char *title, char *filename);
 char GetFileNameLoad(HWND hwnd,char *title, int OpenType=0);
 
-char GetFileNameLoadPtcop(HWND hwnd,char *title);
-
-BOOL LoadPtcopData(void);
