+++ source/PtcopOrg.h
@@ -1,0 +1,2 @@
+char GetFileNameLoadPtcop(HWND hwnd,char *title);
+BOOL LoadPtcopData(void);
